Decouple provider invocation context from runtime
